moved logging and webrtc to require style
