+++ src/freedom/typings/peer-connection.d.ts
@@ -1,7 +1,7 @@
 // Typescript file for core.peer-connection in:
 // https://github.com/freedomjs/freedom/blob/master/interface/core.js
 
-/// <reference path="../../third_party/typings/es6-promise/es6-promise.d.ts" />
+/// <reference path="../../../third_party/typings/es6-promise/es6-promise.d.ts" />
 /// <reference path='freedom.d.ts' />
 
 declare module freedom_PeerConnection {
