+++ src/freedom/typings/tcp-socket.d.ts
@@ -1,7 +1,7 @@
 // Typescript file for core.tcpsocket in:
 // https://github.com/freedomjs/freedom/blob/master/interface/core.js
 
-/// <reference path="../../third_party/typings/es6-promise/es6-promise.d.ts" />
+/// <reference path="../../../third_party/typings/es6-promise/es6-promise.d.ts" />
 
 // The data types used by TcpSocket
 //declare module freedom {
