+++ src/freedom/typings/udp-socket.d.ts
@@ -1,7 +1,7 @@
 // Typescript file for core.udpsocket in:
 // https://github.com/freedomjs/freedom/blob/master/interface/core.js
 
-/// <reference path="../../third_party/typings/es6-promise/es6-promise.d.ts" />
+/// <reference path="../../../third_party/typings/es6-promise/es6-promise.d.ts" />
 
 //declare module freedom {
